:bricks: Separated logic of Client into separate files
